requirements.txt
aerich==0.9.1
aiofiles==24.1.0
aiosmtplib==4.0.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
asyncclick==8.1.8.0
asyncpg==0.30.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
colorama==0.4.6
concurrent-log-handler==0.9.28
cryptography==45.0.5
dictdiffer==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
iso8601==2.1.0
jwt==1.4.0
packaging==25.0
pluggy==1.6.0
portalocker==3.2.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pypika-tortoise==0.6.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
redis==6.4.0
sniffio==1.3.1
starlette==0.47.2
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.35.0
